pyproject.toml
[project]
name = "csv-to-db"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "duckdb>=1.2.1",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "polars>=1.26.0",
    "pyarrow>=19.0.1",
    "pyinstaller>=6.12.0",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
]
